agentproxy/requirements.txt
fastapi
uvicorn
openai
azure-identity
pydantic
